GCS_MAVLink: yaw_sensor is a float

+++ libraries/GCS_MAVLink/GCS_Common.cpp
@@ -3294,7 +3294,7 @@ void GCS_MAVLINK::send_vfr_hud()
         chan,
         vfr_hud_airspeed(),
         ahrs.groundspeed(),
-        (ahrs.yaw_sensor / 100) % 360,
+        ((int32_t)ahrs.yaw_sensor / 100) % 360,
         abs(vfr_hud_throttle()),
         vfr_hud_alt(),
         vfr_hud_climbrate());
